Add files via upload
This is a script of the hand-on session of a tutorial on using Long Short-Term Memory (LSTM) networks for time series forecasting. We demonstrate how LSTM, a type of Recurrent Neural Network (RNN), can model and predict monthly milk production data from 1960 to 1975. Feel free to download it and reuse it.